# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareLens API",
    "description": "Clinical decision-support API for plain radiograph bone tumor detection and Grad-CAM interpretability.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/media/{case_id}/{image_type}": {
      "get": {
        "tags": [
          "Protected Media"
        ],
        "summary": "Get Case Media",
        "description": "Serve uploaded radiograph or Grad-CAM overlay with strict per-case authorization.\nVerifies user ownership (for patients) or clinician role (for doctors) on EVERY image request.",
        "operationId": "get_case_media_media__case_id___image_type__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "image_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Type"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Access token query param for <img> tags",
              "title": "Token"
            },
            "description": "Access token query param for <img> tags"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/{case_id}/{image_type}": {
      "get": {
        "tags": [
          "Protected Media"
        ],
        "summary": "Get Case Media",
        "description": "Serve uploaded radiograph or Grad-CAM overlay with strict per-case authorization.\nVerifies user ownership (for patients) or clinician role (for doctors) on EVERY image request.",
        "operationId": "get_case_media_api_media__case_id___image_type__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "image_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Type"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Access token query param for <img> tags",
              "title": "Token"
            },
            "description": "Access token query param for <img> tags"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}